Added a breach prefix to the session ID
This is in order to mitigate the BREACH SSL attack released at the recent BlackHat conference. I have updated php_send_session_cookie to prefix the session cookie value with a randomly generated prefix (based on the uniqid() function). This prefix is then stripped out of the session ID in php_start_session.